Batch Operations + Utils + Added AudioSample Casting
This commit implements the batch processing functionality of the crate.
The new batch processor (and parallel equivalent) enable library
functions to be applied across entire batches of audio signals.
There is still work to be done to improve this set of features.
- Suring up the API
- Demos
- Benchmarks
- test_convert_fromg

This commit also implements some utility functions such as, but not
limited to:
- correlation
- mse
- mae
- snr
- align_signals

This commit introduces the CastFrom<S> and CastInto<T> traits used to
convert to and from some AudioSample type T and some implementation of
AudioSamples, e.g. for i16

The reasoning behind this funcationality is to allow for operations
such as dividing or multipling or other general operations without
having them be converted into the Audio DSP range. For example, floats
are represented between -1.0 and 1.0. Now imagine you want to divide the
sample by 2.0 (half it). However, AudioSample must divide/mul
AudioSample. So the cast traits allow us to circumvent specific DSP
behaviour and satisfy the compiler.